remove unused OVBaseModelForSeq2SeqLM
